+++ src/main/java/dev/ftb/mods/ftbteambases/config/ServerConfig.java
@@ -16,7 +16,6 @@ import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 
 import java.util.Optional;
-import java.util.function.Consumer;
 
 public interface ServerConfig {
     NameMap<GameType> GAME_TYPE_NAME_MAP = NameMap.of(GameType.ADVENTURE, GameType.values()).create();
@@ -145,43 +144,6 @@ public interface ServerConfig {
         } else {
             FTBTeamBases.LOGGER.error("invalid lobby claim centre pos! expected 2 integers, got {}. default to (0, 0)_", pos.length);
             return new ChunkPos(0, 0);
-        }
-    }
-
-    enum AutoClaimShape {
-        SQUARE("square"),
-        CIRCLE("circle");
-
-        public static final NameMap<AutoClaimShape> NAME_MAP = NameMap.of(SQUARE, values()).id(AutoClaimShape::getId).create();
-
-        private final String shape;
-
-        AutoClaimShape(String id) {
-            this.shape = id;
-        }
-
-        public String getId() {
-            return shape;
-        }
-
-        public void forEachChunk(ChunkPos origin, int radius, Consumer<ChunkPos> consumer) {
-            BlockPos pos0 = origin.getMiddleBlockPosition(0);
-            int blockRadiusSq = (radius << 4) * (radius << 4);
-            switch (radius) {
-                case 0 -> { }
-                case 1 -> consumer.accept(origin);
-                default -> {
-                    int r = radius - 1;
-                    for (int cx = origin.x - r; cx <= origin.x + r; cx++) {
-                        for (int cz = origin.z - r; cz <= origin.z + r; cz++) {
-                            ChunkPos cp = new ChunkPos(cx, cz);
-                            if (this == SQUARE || cp.getMiddleBlockPosition(0).distSqr(pos0) < blockRadiusSq) {
-                                consumer.accept(cp);
-                            }
-                        }
-                    }
-                }
-            };
         }
     }
